Simplify Makefile patsubst function
Instead of having:

    SRCFILES= $(patsubst %.o,%.c,$(OBJFILES))
    DSYMDIRS= $(patsubst %,%.dSYM,$(TARGETS))

We now have:

    SRCFILES= $(OBJFILES:.o=.c)
    DSYMDIRS= $(TARGETS:=.dSYM)

The old way is commented out noting another way to do it. The reason I
have done this is I noticed it when reading about the patsubst function
(I'm the one who actually added the SRCFILES variable but I was looking
for something else) and it seemed like a worthy change. Works under
macOS and linux (checked CentOS 7 so should be fine in others).

There's a specific reason I was encountered this but if I make use of it
(might do as it'll be a nice addition to the Makefile) it'll be in a
future commit.