Makefile updates wrt man pages and install
The MANPAGES variable is now calculated via the new variable
MAN_TARGETS. Since probably all the TARGETS should eventually have a man
page the assignment of MANPAGES should eventually work on TARGETS but
until then we use the MAN_TARGETS variable.

As well I've added the -v option to the install invocations in make
install to show the user where the files were in fact installed (even if
it might be obvious to most from the last argument to each invocation).

Notes on the man pages in the Makefile:

    # Currently we explicitly specify a variable for man page targets as not all
    # targets have a man page but probably all targets should have a man page. When
    # this is done we can change the below two uncommented lines to be just:
    #
    #   MANPAGES = $(TARGETS:=.1)
    #
    # But until then it must be the next two lines (alternatively we could
    # explicitly specify the man pages but this makes it simpler). When a new man
    # page is written the MAN_TARGETS should have the tool name (without any
    # extension) added to it.  Eventually MAN_TARGETS can be removed entirely and
    # MANPAGES will act on TARGETS.

    MAN_TARGETS = mkiocccentry txzchk fnamchk iocccsize jinfochk jauthchk
    MANPAGES= $(MAN_TARGETS:=.1)

This change is what I referred to in commit 6aef3c (I was not sure of
the part without an extension though I noticed it was right in front of
me in the Makefile though I still think the simplification is worth it).