Typo fixes in jstr-test.sh
I saw this when the script was first introduced but I couldn't be sure
if it was a typo or not. But since I can't figure out what 'foles' means
I can only think it was supposed to be 'holes'. If this is not correct
I"m happy to correct it to what it should be even if that's 'foles'.