Fix compilation under CentOS 7
The function floorl() requires -lm under CentOS 7 (the proper feature
test macros were already enabled although I have also changed
_XOPEN_SOURCE to be 600 up from 500: not sure if this is strictly
necessary but it doesn't hurt to have it at 600).

The Makefile has had some changes in the way warnings are handled
described below as thus:

First the warning flags -Wall -Wextra -Werror have been moved to the new
variable WARN_FLAGS. After this in WARN_FLAGS specific warnings have
been disabled in particular:

     -Wno-unused-command-line-argument -Wno-poison-system-directories -Wno-unreachable-code-break -Wno-padded

The first one is important for macOS as after adding -lm to the CFLAGS
(via the standard LDFLAGS variable) we get that error for the fewer
objects that don't use -lm. Since many of the objects do link in json.o
(which is what needs -lm) we simply add disable the warning in the
WARN_FLAGS variable. There is at least one warning that we explicitly
disable in some rules as they're not needed in many: for example the
parser needs some warnings disabled.

As for the others they're enabled by -Weverything and although we don't
enable that they're here in case we ever do as they're irrelevant here.

With this commit everything should compile under macOS and also CentOS 7
(and other linux systems should be fine as well: CentOS is in a sense
behind as it's built for stability). In 2024 CentOS 7 will reach EOL and
some of the things we include for CentOS 7 should thus no longer be
needed.