English updates: 1st -> first, 2nd -> second etc.
I have changed 1st -> first, 2nd -> second etc.

I've also changed 2 args required to two args required. This one is
questionable in technical terms but for now it'll do. I've not looked
for more than '2 args' or '2 arguments' or other variations. For example
I'm aware of:

    dbg.c:	fprintf_usage(DO_NOT_EXIT, stderr, "requires 2 or 3 arguments");

and there are probably others like it but this can come at another time
(either revert back to [0-9] args/arguments or change them all to words;
either way numbers > 9 should be digits).

As some other changes are imminent with json.c and since this update
does in fact touch json.c this commit can be rejected. All one has to do
to fix this (less the '2 args' part) is to use a sed that supports -i
(e.g. GNU sed) and do:

    sed -i -e '1st/first/g' -e 's/2nd/second/g' -e 's/3rd/third/g' -e \
    's/4th/fourth/g' -e 's/5th/fifth/g' -e 's/6th/sixth/g' *.[ch]

The reason I did not go beyond 6th is because that appeared to be the
highest number.

Caveat with the sed invocation above: I typed this out in the commit log
and I did it the lazy way at the command line (one at a time) so if I
made a mistake that's why but the idea is what matters.