Typo fix and digit -> word changes
In jint.h it referred to the earlier named struct Integer (is struct
integer).

Also text in the form of 'requires 2 or 3 arguments' are now 'requires
two or three arguments'. There are some files I have left it out though:
in scripts that have text like: 'expected 2 arguments, found $#' I kept
the 2 in because the $# will also consist of digits and consistency is
important.