More [0-9] args/arguments <-> word changes + typos
In jfloat.h:

    -    "\tNOTE: Without -t, at least 1 argument is required.\n"
    +    "\tNOTE: Without -t, at least one argument is required.\n"

In vermod.sh change the word to a number in the same logic as in commit
b0d291: since the $# shows digit(s) not a word the other number should
also not be a word but (a) digit(s).

In util.c there was a typo fix and also:

    - *	    .. data has at least 1 internal NUL byte ..
    + *	    .. data has at least one internal NUL byte ..

In json.c I changed 1 -> one and 5 -> five.

There very possibly are other places where there is something like this
but for args/arguments all instances should be updated and consistent in
form.