Remove useless code + bug fix author handle prompt
For each function there's a do while loop with the condition that the
data has to be invalid in order for the loop to continue (i.e. as long
as it's NULL or out of range the loop continues). But after each loop
there was a check for the invalid values that the loop makes sure is not
invalid with (if it was invalid which it never would be except for the
one bug below that's also now fixed) an error message that said "retry
prompt disabled" which one would have to look at the code to figure out
what happened rather than get a descriptive error message. This actually
is a completion of commit 915bf7cb and in particular this part:

    Removed need_retry variable as it appeared to have no effect on
    code.

In making this change I also discovered a bug in validating author
handles: when posix_plus_safe() returned false there was a break
statement rather than a continue (which technically is not needed as
it's the final check in the loop but in case anything else is ever added
after it I changed it to continue) which means that the above check that
I removed have had a use but it shouldn't have been reached at all.

Sequenced exit codes.