Fix commit b0813e09
The change was not done in jfloat.h or jint.h and I'm not sure why as I
was pretty sure I used sed -i on *.h. However I've been having some
issues with my fork today and when making this commit I noticed some
additional things that are not right as well. Anyway this commit changes
usage_msgX to be usage_msg([0-9]?)+ for the reasons I cited in the other
commit.