JSON scanner and parser updates
We now #include "jparse.tab.h" in jparse.h. The solution to the previous
problem (where the file could not be included in jparse.h) was to #define
YYDEBUG in jparse.h prior to the inclusion of jparse.tab.h.

The following changes were made in the scanner:

* Whitespace is now printed in the scanner so as to distinguish it from
other characters. We do NOT return a type for it because this would
screw up and confuse the parser (more than it already is screwed up and
confused that is :) ).

* Point out in comments that the JTYPE_STRING regex is incorrect. Also
added a simple regex (in comments) for reference of what a string would
look like without the JSON rules. Once JTYPE_STRING is correct this
comment will be removed.

* Added JTYPE_HEX regex which is:

    u[0-9A-Fa-f]{4}

The reason there's no \\ before the 'u' is because in JTYPE_STRING the
\\ is already there as it's also used for \- escape characters. It's not
yet used in the JTYPE_STRING regex but will be in time to simplify the
regex for easier human parsing.

* Added JTYPE_UNICODE. This set is a negated character set and it's not
used yet and is possibly incorrect. It might or might not be used but it
probably will be in time, possibly modified (and if necessary
corrected), to break down the string regex for easier human parsing.

* Add some additional comments.

* Add newline in printf statement for each regex match prior to printing
the type.

* Add debug output in the form of

    *** BEGIN DEBUG

    '
    <string>
    '

    *** END DEBUG

to show what it's about to parse (or as is the current case what it's
about to attempt to parse).

The following changes were made in the parser:

* Add debug output (debug level DBG_NONE so it's always printed - at
least for now) saying that it's calling parse_json_file() or
parse_json_string() with the respective parameter. This dbg() call will
remain in place but the debug level will be changed later on.