Minor Makefile improvement
Add to make prep and make build a message telling the user where they
can find the log of all the checks. In the Makefile it looks like:

	@echo NOTE: The above details were saved to ${BUILD_LOG}.

which by default would print:

	NOTE: The above details were saved to build.log.

This is done regardless of the exit status of ./prep.sh.

I also aliased make build to release and pull as it's meant to be run
prior to a pull request (though some checks don't have to pass
necessarily).

There were some comments updates as well:

thru -> through (similar logic as when I did this in other files).

Some longer lines were split into two lines.