flex and bison errors are now fatal
In order to prevent accidental commit of invalid *.[ly] files (which can
easily happen if it doesn't result in a compilation error) the Makefile
now does not proceed if flex or bison fails, exiting on the first error.
For example if flex cannot generate the appropriate files you'll see
an along the lines of:

    unable to form jparse.c
    if you've not modified jparse.l you can run:

	    make use_ref all

    to use the backup file. If you do this please
    consider reporting this as a bug. If you've
    modified jparse.l please fix it and rerun this rule.
    make: *** [jparse.c] Error 1

    shell returned 2

Similarly if bison fails to generate the appropriate files you'll see:

    unable to form jparse.tab.h and/or jparse.tab.c
    if you've not modified jparse.y you can run:

	    make use_ref all

    to use the backup files. If you do this please
    consider reporting this as a bug. If you've
    modified jparse.y please fix it and rerun this rule.
    make: *** [jparse.tab.h] Error 1

    shell returned 2

though you might also see a compiler error.

This is safe because the bfok.sh script verifies that the versions are
correct. If more bison or flex files are created these rules will have
to be modified. It might also be an idea to check the exit status of
these commands rather than just checking for 0 length files.