Major (mostly JSON) reorganisation
Move JSON parser specific structs and functions from json.{h,c} to
jparsefuncs.{h,c} with the exception of struct json, its union (and the
related structs) and code that acts on these. Perhaps these should also
be moved over and the files renamed jparseutil (or kept as jparsefuncs
which looks nicer to my eyes though possibly it's not as accurately
named: not sure) or something like that?

Move the functions in mkiocccentry.{h,c} that write a json file to
json.{h,c}. Perhaps these could be in jutil.{h,c} (which was the
original name thought up for jparsefuncs.{h,c}) along with the
encode/decode functions: not sure.

Moved lookup_location_name() to location.c where it belongs. This is the
only change not specific to JSON.

Updated Makefile (including make depend) for these changes.

I do want to point out some intentional irony in the comments at the top
of jparsefuncs.{h,c}. It has the quote that I added for txzchk:

    + * "Because sometimes even the IOCCC Judges need some help." :-)

but below that I put:

    + * This is currently being worked on by:
    + *
    + *	@xexyl
    + *	https://xexyl.net		Cody Boone Ferguson
    + *	https://ioccc.xexyl.net
    + *
    + * and
    + *
    + *	chongo (Landon Curt Noll, http://www.isthe.com/chongo/index.html) /\oo/\

when of course Landon is one of the judges ... which means that the judges
are getting help and they're also helping themselves by helping someone
help them! :)

TODO:

* Decide whether or not to move the struct json and the related structs
and functions to jparsefuncs. This might seem like the obvious choice
but the discussion did not originally suggest this so I'm keeping it in
json.{h,c} for now.

* Decide if anything else should be moved over.

* Decide if jparsefuncs is the best name for the above mentioned
changes.