Simplify JTYPE_STRING regex
Revert back to the much less complicated regex. The following note is in
a comment above it as it does not yet handle everything correctly (and
it never did):

    * Now one might ask questions about the tighter restrictions on JSON strings
    * and why we don't even consider them. This is a good question but the answer
    * is simple: once more is done with the parser structs and functions we will
    * let the conversion routines do the actual JSON checks: this way it keeps the
    * regex simple and we don't have to worry about any complexities.

For this change I've removed JTYPE_UNICODE and JTYPE_HEX both of which
should not be needed (and based on some discussions after re-evaluating
the spec the other day the hex one seemed like it might be incorrect
anyway).