Tokens and symbols renamed: JSON_ <-> JTYPE_
As discussed on GitHub issue about JSON (see
https://github.com/ioccc-src/mkiocccentry/issues/156#issuecomment-1105550027)
I've made the grammar for flex/bison use JSON_ instead of JTYPE_ so that
it looks more like the spec of JSON (conversely, of course, the previous
JSON_ symbols were renamed to be JTYPE_).

To make the lines in the rules of jparse.y look a bit more aligned I
have the '|' at the beginning of each line. I don't actually like it
this way as much (yet) but I think it results in cleaner line endings as
the '|' is no longer at different columns.