Older bison and POSIX Yacc compatibility
Change parse.error 'detailed' to 'verbose' to support older versions of
bison. Now bison version 3.0.4 will work which means CentOS 7 bison can
actually generate the file. The change in verbosity does not seem to
make a difference for me so this is a worthwhile change. Please note the
version has not yet been updated in version.h.

Now passes -Wyacc which has been added to the Makefile. This is done by
making a couple changes described below.

POSIX Yacc does not support the %empty directive so we use /* empty */
instead.

POSIX Yacc does not support the %define directive but it does support
the -D option. Thus instead of having in jparse.y (where detailed is now
verbose as described above):

    %define parse.error detailed
    %define parse.lac full

we have in the Makefile:

    echo "$$BISON_PATH -d -Wyacc -Dparse.error=verbose -Dparse.lac=full jparse.y"; \
    "$$BISON_PATH" -d -Wyacc -Dparse.error=verbose -Dparse.lac=full jparse.y; \