Minor bison rules update
Add json_number which has the three types of numbers (to be changed to
either one token or two: integer and float or just number). This is
instead of having the three types on a line each under json_value.
Specifically the following diff was made:

    -json_value:    json_object
    +json_value:    json_object
		    | json_array
		    | JSON_STRING
    -		    | JSON_INTMAX
    -		    | JSON_UINTMAX
    -		    | JSON_LONG_DOUBLE
    +		    | json_number
		    | JSON_TRUE
		    | JSON_FALSE
		    | JSON_NULL
		    ;

    +json_number:   JSON_INTMAX
    +		    | JSON_UINTMAX
    +		    | JSON_LONG_DOUBLE ;
    +

I've also updated the comments above these suggesting that the grammar
is more or less complete except for the numbers and the actions for each
rule though it might be that I have indeed missed something else.