JSON warning and error updates
The json_err.codes and json_warn.codes files have been reworded
particularly in the description of the format of the codes and (for
json_err.codes) use __FILE__ in place of the filename to account for the
possibility that the file changes which happened recently.

In check_common_json_fields_table() (json_chk.c) I added 'at' before the
word 'location' like it is for the other errors in that form.

Added a note in jerr() and jwarn() (and their errno counterparts)
functions that sometimes the file is the source file that raised the
issue.