verge: malloc_ -> allocate_
Since the function uses calloc() I have changed the name to account for
this. I did not choose calloc_ because I feel that allocate_ (alloc_
would also work) is more descriptive (even if calloc() is familiar to
anyone who would be participating in the IOCCC).

There are still many places where malloc() is used and there might be
places where calloc() is used but comments refer to malloc(). I just
noticed this one by chance and decided to fix it.

There was a bit of rewording in verge.c as well.