Improve JSON errors - use literal string tokens
This will provide for better error messages. For example instead of
seeing something like:

    JSON parser error (num errors: 1) on line 1: syntax error, unexpected JSON_CLOSE_BRACE

one will see:

    JSON parser error (num errors: 1) on line 1: syntax error, unexpected }

which is much easier to read.

This is not POSIX Yacc portable but we deem that the benefits outweigh
this issue. Because of this we also make use of the %define directive to
make it easier to find (all the comments, many of which will be removed
later on, notwithstanding of course) in the grammar itself (this
directive is also not POSIX Yacc portable but -D is).