Makefile dependency updates for flex and bison
If sorry.tm.ca.h or jparse.h is updated both flex and bison will
regenerate the files. This is because if jparse.h is updated it might be
a change in prototypes (or new functions or variables or something else)
and if sorry.tm.ca.h is updated we want the updated apology to be
injected into the files again.