Fixed make build release pull to not use backup files
The build release pull will run prep.sh with -o,
which in turn will run run_bison.sh and run_flex.sh with -o.

The effect of -o is to prevent the bison and flex backup files
from being used.  If -o is given and bison and/or flex is
not found or is too old, then the operation will fail with
a non-zero exit code.

The make prep rule will, and before, make use of
backup files, unless the RUN_O_FLAG is set to -o
(which it is not by default).