Lexer and parser ironically declare they're ugly
This is done by changing the (yy|YY) prefix to (ugly|UGLY) in the lexer
and parser. As the following comment says:

    /*
     * As we utterly object to the hideous code that bison and flex generate we
     * point it out in an ironic way by changing the prefix yy to ugly_.
     */

This is done in flex and bison via:

    %option prefix="ugly_"
    %define api.prefix {ugly_}

respectively. This creates an ironic statement in the lexer:

    #define ugly_lex_ALREADY_DEFINED

and an ironic statement in the C pre-processor in the parser:

    #if ! defined UGLY_STYPE && ! defined UGLY_STYPE_IS_DECLARED
    # define UGLY_STYPE_IS_DECLARED 1

which is to say that both the lexer and parser actually declare that
they're ugly! :))

For the curious:

    $ grep -ic ugly jparse*ref.[ch]
    jparse.ref.c:108
    jparse.tab.ref.c:34
    jparse.tab.ref.h:24

which is a huge amount of ugliness (though it's much more ugly than what
it admits :) )!

As well I've made a minor change to sorry.tm.ca.h that's not even worth
describing.