Use our struct json instead of bison %union
This is accomplished in the parser by:

    %define api.value.type {struct json}

in the parser. There's a minor uncomfortable fact though with the way we
prefix the lexer/parser with ugly_ instead of yy:

    typedef struct json UGLY_STYPE;

which is of course suggesting that our struct json is ugly when actually
it's bison and flex that is ugly. I'm not sure if this is a problem or
not. If it is we'll possibly undo that irony but for now it's there.

As well I've made use of the union within struct json in the lexer for
the types that exist (there are no types for e.g. '{', '}', ':' etc.) or
at least set the type of the member. Whether how this is done changes or
not I do not know yet.