txzchk: update total issues in cases missed
There were several places where warn() is called but
txz_info.total_issues was not incremented. For example when encountering
a NULL pointer in parsing. It might be that it could be a different tar
output format (though in some cases this wouldn't be the case maybe) but
whether it's that or something else it should be noted as a warning (as
it was) but also the number of issues should be incremented so that it
doesn't somehow pass all tests.

Note that not all calls to warn() warrant a problem with the tarball.
There are some places where the issues were already incremented but
now we give an overall summary. There are other places where warn() does
not imply an issue itself.

A point of interest perhaps is that in some places warn() is called
before incrementing the total number of issues and in other places it's
called after. I don't like this inconsistency but because everything
works I will not fix it until a test suite has been built for it (I
added the features in the first place that allow for this to be done but
other things have higher priority as it works well).