Improve JSON definition + parser comments
Remove some outdated comments and add some comments about what needs to
be done, what is done etc.

As well I've improved the redefinition of JSON when referring to the
ugly_ prefix. In particular:

    - * surely be in the JSON Barmy Army (otherwise known as the Barmy Army of the
    - * Jocular Sillies On Narcotics :-)).
    + * surely be in the JSON Barmy Army (otherwise known as the Barmy Army Jointly
    + * Staying On Narcotics :-)).

And to those who do not know and might be curious: barmy army are
supporters of a particular sports team known for vociferous support
during matches especially fans of the England cricket team. Barmy (which
I also referred to in an earlier commit) is a term (also British
English) which means mad, crazy; extremely foolish.

Thus I was making a number of puns when likening the supporters of the
JSON standard as being in the obviously made up JSON Barmy Army but all
in good fun.