Scripts not needing man page now in SH_TARGETS
The eventual plan is that the man pages variable will rely on TARGETS
but there is no need for a man page for limit_ioccc.sh (but see below).

There might be some shell scripts (e.g. run_flex.sh and run_bison.sh)
that could use a man page but these can be left in TARGETS at that point
(if they're deemed worthy of being installed). Or it might be that the
man pages variable works on more than one target variable. But for now
the man page targets are in their own variable (until such a time that
all targets have a man page).

The idea of the man pages relying on the TARGETS variable could however
cause a problem if ever a tool is written without an accompanying man
page. On the other hand it might encourage the person who writes the
tool to write a man page as well. Perhaps also a better variable name
could be contrived given that some shell scripts might have a use for a
man page but for now this will suffice.