Lexer and parser updates
In jparse.l specify option 8bit. The Makefile still passes -8 to
run_flex.sh but we have both in case one or the other is removed.

Make full use of the ugly_ prefix (but see below). That is change all
references in the scanner and parser to refer to ugly_ instead of yy_
(we don't use any of the YY - yet at least).

Noted that the ugly_ prefix is a satire for the IOCCC and if it's moved
to a new repo for a JSON parser (as was suggested to me) after the IOCCC
has used it I might well remove the prefix even though we all think the
generated code by flex and bison is in fact ugly. Perhaps this should be
moved further up in the comment but this can come at another time.

Add comments to the idea of (ugly_|UGLY_). Specifically I've noted how
to see what was (re?)defined to something ugly as well as how to see
only symbols changed to something ugly. This might also be of use should
one need to find out what name should be used.