Add jparse.1 and add DBG_VERSION to version.h
MAN_TARGETS in Makefile updated so that make install will install
jparse.1.

VERSION in dbg.c itself -> DBG_VERSION in version.h. This allows for the
version to be in limit_ioccc.sh.

 -V added to dbg to print version.

The change to dbg necessitated make depend as well as the dbg.h was
updated to #include "version.h".