Add warning to lexer/parser prefix comments
This warning warns against trying to sift through the flex and bison
generated code but also reminds them that if they do choose to go
through that code the prefix will at times be yy and YY and other times
be ugly_ and UGLY_ (part of this is because we made it do so as satire
for the IOCCC). This warning also refers back to sorry.tm.ca.h.