Clean up in test_JSON (mostly timestamp specific)
Many files in the test_JSON tree had invalid timestamps as the
MIN_TIMESTAMP was recently changed. I reset the timestamp in the files
to be the value of MIN_TIMESTAMP with the exception of the files with
names that suggest that they should be wrong. Specifically I did not
change:

    test_JSON//info.json/bad/info.timestamp-2.json
    test_JSON//info.json/bad/info.timestamp-1.json
    test_JSON//info.json/bad/info.timestamp-0.json

I also renamed 'test_JSON/author.json/bad/author.version-4.json' ->
'test_JSON/author.json/bad/mkiocccentry.version-4.json' as it appears
that the invalid version is actually mkiocccentry as it has unescaped
'/'s in it. I am not clear what's wrong with the other ones that have
author.version in the filename as it appears that the version is correct
at least for author version.