Change some NOTE comments to XXX comments
Note that the check_(first|last)_json_char() functions will be removed
in time but we're keeping them in for now as it would require changing
the jinfochk and jauthchk tools temporarily only so they can work (in so
far as they are done) for the mkiocccentry generated files only to have
to be removed later when the parser is complete. It's better to work on
things (including the parser) that need to be done rather than something
that will be removed later on.