Improve txzchk exit codes and update txzchk.1
Now only exits 0 if the tarball was successfully parsed and there were
no issues. Previously printing the usage or version would also exit with
0 and the only other times it would exit non-zero is if there was an
error.

usage() now exits with exit code 2.

Add missing seqcexit comment /*ooo*/ to a call to the usage() function.

Rewrite exit status section in txzchk.1 to make it clearer and more
correct (it was actually incorrect even before the changes to the code).

Only warn when welcome string printf fails (previously would abort when
this is not so important).