Update -C to -T in mkiocccentry + better errors
In mkiocccentry the option to specify the path to txzchk was -C but it
should have been -T (-t not possible as that's the option for path to
tar).

In several files where the call to is_file() returned false the error
message would say it's not a file but it actually tests for it being a
regular file and it is confusing to say to the user it's not a file so
now it says it's not a regular file (there were some places where I
already did this but I did not do it everywhere).