Add #define JPARSE_PATH_[01] to util.h
Currently this is not needed but since it's our tool and it might in
time be used directly by another tool I've added it to. The find_utils()
function does not act on it yet as that would require updating all calls
to it and it's not necessary at this time.