JSON parser updates
Fix jparse.1 SYNOPSIS section. It suggested it will parse a string
without an option and did not suggest anything to do with a file.

Fix jparse main() to not check for option 'S' which is the invalid idea
of strictness.

Add start of parse functions for the json types: member, array, string,
name (this will be a wrapper function), number, boolean and null. These
are incomplete and subject to change. All they do now is check for NULL
pointers and then return a NULL struct json *.

Added some XXX comments suggesting that some functions do not belong in
the jparse.l and jparse.y files and this should be addressed in the
future (there were other XXX comments added to the new functions as
well).