Add conversion code to most parse_json_ functions
For most of these (I believe the sole exception being boolean) there
still are no tests being done on the conversion: just check for NULL
pointer. For boolean we check if it converted because if not it suggests
a serious mismatch: only booleans should be passed to this function and
that means "true" or "false" and nothing else. Thus if conversion is
false it's an error and the program will be aborted.

TODO:

  The array and member functions do not yet parse anything. I believe
  everything else does.

  Nothing is linked into the tree yet and neither are these functions
  called in the parser.

  Decide what tests should be done on conversions and what actions
  should be taken. It's likely that a conversion error in JSON strings
  should be a fatal error (and it might be indirectly - not sure) for
  example. This goes for a number of types.

There was a typo fix in json.c as well (or rather: surround a " with 's
in comments for easier reading in syntax highlighting).